+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rc-checkbox",
-  "version": "2.1.8",
+  "version": "2.2.0",
   "description": "checkbox ui component for react",
   "keywords": [
     "react",
@@ -54,9 +54,7 @@
   },
   "dependencies": {
     "babel-runtime": "^6.23.0",
-    "classnames": "2.x",
-    "prop-types": "15.x",
-    "react-lifecycles-compat": "^3.0.4"
+    "classnames": "2.x"
   },
   "devDependencies": {
     "core-js": "^2.5.1",
